+++ Release-Notes.md
@@ -1,5 +1,18 @@
 # FarNet.Redis Release Notes
 
+## v0.5.0
+
+Cmdlets value parameters use `object` instead of `string` and support `byte[]` values.
+Output still uses strings. Use SERedis methods to get `byte[]` results, e.g.:
+
+    [byte[]]$db.StringGet(..)
+
+See test tasks `bytes` for examples of using `byte[]`.
+
+`Get-RedisString`, potentially breaking change
+- `-Key` is `string`, not `string[]`
+- use `-Many` for getting 2+ strings
+
 ## v0.4.0
 
 Potentially breaking changes in cmdlet parameters.